fix: Move Sequence type to data_types2
