+++ write_buffer/src/kafka/mod.rs
@@ -10,7 +10,8 @@ use crate::{
     },
 };
 use async_trait::async_trait;
-use data_types::{sequence::Sequence, write_buffer::WriteBufferCreationConfig};
+use data_types::write_buffer::WriteBufferCreationConfig;
+use data_types2::Sequence;
 use dml::{DmlMeta, DmlOperation};
 use futures::{stream::BoxStream, StreamExt};
 use iox_time::{Time, TimeProvider};
