+++ write_summary/src/lib.rs
@@ -1,14 +1,11 @@
-use std::collections::BTreeMap;
-
 use data_types2::{KafkaPartition, SequenceNumber};
-use observability_deps::tracing::debug;
-
+use dml::DmlMeta;
 /// Protobuf to/from conversion
 use generated_types::influxdata::iox::write_summary::v1 as proto;
-
-use dml::DmlMeta;
-
+use observability_deps::tracing::debug;
 use snafu::{OptionExt, Snafu};
+use std::collections::BTreeMap;
+
 mod progress;
 pub use progress::SequencerProgress;
 
@@ -203,7 +200,7 @@ impl TryFrom<proto::WriteSummary> for WriteSummary {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use data_types::sequence::Sequence;
+    use data_types2::Sequence;
 
     #[test]
     fn empty() {
